requirements.txt
awscli==1.16.74
boto3==1.9.71
botocore==1.12.71
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
Django==2.1.1
django-datetime-widget==0.9.3
django-filter==2.0.0
djangorestframework==3.9.0
docutils==0.14
ez-setup==0.9
future==0.17.1
idna==2.8
jmespath==0.9.3
jsonfield==2.0.2
munch==2.3.2
numpy==1.15.4
psycopg2==2.7.6.1
pyasn1==0.4.4
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
serializers==0.2.4
six==1.12.0
urllib3==1.24.1
